Fixed Data Types
For some reason the "Decimal" does not exist (on Mac) despite its presence in the docs: https://learn.microsoft.com/office/vba/language/reference/user-interface-help/data-type-summary#set-intrinsic-data-types

As such, we resort to the "Double" which should always be large enough to hold the ".CountLarge" value: https://learn.microsoft.com/office/vba/api/excel.range.countlarge#remarks

+++ src/sPrinter.bas
@@ -1136,8 +1136,8 @@ Private Sub CheckRange( _
 	
 	
 	' Measure the dimensions of the Range.
-	Dim nRows As Decimal: nRows = rng.Rows.CountLarge
-	Dim nCols As Decimal: nCols = rng.Columns.CountLarge
+	Dim nRows As Double: nRows = rng.Rows.CountLarge
+	Dim nCols As Double: nCols = rng.Columns.CountLarge
 	
 	' Throw an error for a rectangular (2D) area.
 	If nRows > 1 And nCols > 1 Then
